+++ common/src/config/GameConfig.cpp
@@ -218,6 +218,7 @@ bool GameConfig::visit_vars(T&& visitor, bool is_save)
     result &= visitor(dash_faction_key, "Already Saw First Launch Window", suppress_first_launch_window);
     result &= visitor(dash_faction_key, "Multi Ricochet", multi_ricochet);
     result &= visitor(dash_faction_key, "Unlimited Semi Auto", unlimited_semi_auto);
+    result &= visitor(dash_faction_key, "Use Gaussian Bullet Spread", gaussian_spread);
     result &= visitor(dash_faction_key, "World HUD CTF", world_hud_ctf);
     result &= visitor(dash_faction_key, "World HUD Overdraw", world_hud_overdraw);